+++ rackspace_monitoring/drivers/rackspace.py
@@ -184,10 +184,19 @@ class RackspaceMonitoringDriver(MonitoringDriver, OpenStackDriverMixin):
 
         if ex_force_base_url and not ex_force_auth_token:
             self.connection._populate_hosts_and_request_paths()
-            ep = self.connection.service_catalog \
-                     .get_endpoint(name=self.connection.service_name,
-                                   service_type=self.connection.service_type,
-                                   region=None)
+            eps = self.connection.service_catalog \
+                      .get_endpoints(name=self.connection.service_name,
+                                     service_type=self.connection.service_type)
+
+            if len(eps) == 0:
+                raise Exception('Service catalog contains no endpoints')
+
+            # Cloud Monitoring only has single endpoint so use a first available
+            # one.
+            # Note: We cannot use "get_endpoint" because the staging auth
+            # catalog entry contains a "region" attribute which we don't know
+            # up front.
+            ep = eps[0]
 
             tenant_id = ep['tenantId']
             self.connection._ex_force_base_url = '%s/%s' % (