+++ CHANGES.md
@@ -1,3 +1,8 @@
+rackspace-monitoring in development
+
+* Fix a bug with service catalog parsing which would occour when user uses
+  `ex_force_base_url` argument with the staging auth endpoint.
+
 rackspace-monitoring v0.5.0 - 2013-06-06:
 
 * Use service catalog to determine the API endpoint url.